Singular reward class
